% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-classes.R
\name{Boolean}
\alias{Boolean}
\title{Boolean}
\description{
A boolean field
}
\keyword{internal}
\section{Super classes}{
\code{\link[staypuft:FieldABC]{staypuft::FieldABC}} -> \code{\link[staypuft:Field]{staypuft::Field}} -> \code{Boolean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Boolean$new()}}
\item \href{#method-serialize_}{\code{Boolean$serialize_()}}
\item \href{#method-deserialize_}{\code{Boolean$deserialize_()}}
\item \href{#method-clone}{\code{Boolean$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="bind_to_schema">}\href{../../staypuft/html/Field.html#method-bind_to_schema}{\code{staypuft::Field$bind_to_schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="context">}\href{../../staypuft/html/Field.html#method-context}{\code{staypuft::Field$context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="deserialize">}\href{../../staypuft/html/Field.html#method-deserialize}{\code{staypuft::Field$deserialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="fail">}\href{../../staypuft/html/Field.html#method-fail}{\code{staypuft::Field$fail()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="get_value">}\href{../../staypuft/html/Field.html#method-get_value}{\code{staypuft::Field$get_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="print">}\href{../../staypuft/html/Field.html#method-print}{\code{staypuft::Field$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="root">}\href{../../staypuft/html/Field.html#method-root}{\code{staypuft::Field$root()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="serialize">}\href{../../staypuft/html/Field.html#method-serialize}{\code{staypuft::Field$serialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="validate_">}\href{../../staypuft/html/Field.html#method-validate_}{\code{staypuft::Field$validate_()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="validate_missing_">}\href{../../staypuft/html/Field.html#method-validate_missing_}{\code{staypuft::Field$validate_missing_()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Boolean object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boolean$new(..., truthy = NULL, falsy = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{truthy}}{Values that will (de)serialize to \code{TRUE}. If an
empty set, any non-falsy value will deserialize to \code{TRUE}. If \code{NULL},
xx will be used.}

\item{\code{falsy}}{Values that will (de)serialize to \code{FALSE}. If \code{NULL},
xx will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize_"></a>}}
\if{latex}{\out{\hypertarget{method-serialize_}{}}}
\subsection{Method \code{serialize_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boolean$serialize_(value, attr = NULL, obj = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserialize_"></a>}}
\if{latex}{\out{\hypertarget{method-deserialize_}{}}}
\subsection{Method \code{deserialize_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boolean$deserialize_(value, attr = NULL, data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boolean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
