% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_registry.R
\name{ClassRegistry}
\alias{ClassRegistry}
\title{ClassRegistry}
\description{
Base schema class with which to define custom schemas
}
\examples{
x <- ClassRegistry$new()
x
x$registry
z <- Schema$new("FooBar",
  name = fields$character(),
  title = fields$character()
)
w <- Schema$new("MySchema",
  name = fields$character(),
  title = fields$character()
)
x
x$registry
x$register("FooBar", z)
x$register("MySchema", w)
x
x$registry

x$get_class("FooBar")
x$get_class("MySchema")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{registry}}{(list) list of classes registered}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{ClassRegistry$print()}}
\item \href{#method-register}{\code{ClassRegistry$register()}}
\item \href{#method-get_class}{\code{ClassRegistry$get_class()}}
\item \href{#method-clone}{\code{ClassRegistry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassRegistry$print(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register"></a>}}
\if{latex}{\out{\hypertarget{method-register}{}}}
\subsection{Method \code{register()}}{
Add a class to the registry of serializer classes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassRegistry$register(class_name, cls)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class_name}}{(character) the class name}

\item{\code{cls}}{(Schema) the class object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing, registers internally
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_class"></a>}}
\if{latex}{\out{\hypertarget{method-get_class}{}}}
\subsection{Method \code{get_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassRegistry$get_class(class_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassRegistry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
