% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_field_schema.R
\name{FieldABC}
\alias{FieldABC}
\title{FieldABC}
\description{
Abstract base class from which all Field classes inherit
}
\examples{
\dontrun{
x <- FieldABC$new()
x
x$serialize()
x$deserialize()
x$serialize_()
x$deserialize_()
}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parent}}{xxx}

\item{\code{name}}{xxx}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-serialize}{\code{FieldABC$serialize()}}
\item \href{#method-deserialize}{\code{FieldABC$deserialize()}}
\item \href{#method-serialize_}{\code{FieldABC$serialize_()}}
\item \href{#method-deserialize_}{\code{FieldABC$deserialize_()}}
\item \href{#method-clone}{\code{FieldABC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\subsection{Method \code{serialize()}}{
serialize fun: to be replaced by child class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldABC$serialize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserialize"></a>}}
\subsection{Method \code{deserialize()}}{
deserialize fun: to be replaced by child class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldABC$deserialize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize_"></a>}}
\subsection{Method \code{serialize_()}}{
serialize_ fun: to be replaced by child class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldABC$serialize_()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserialize_"></a>}}
\subsection{Method \code{deserialize_()}}{
deserialize_ fun: to be replaced by child class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldABC$deserialize_()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldABC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
