% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\docType{data}
\name{fields}
\alias{fields}
\title{staypuft fields}
\format{
An object of class \code{list} of length 11.
}
\usage{
fields
}
\description{
staypuft fields
}
\section{Types of fields}{

\itemize{
\item \link{Missing}
\item \link{Character}
\item \link{UUID}
\item \link{Number}
\item \link{Integer}
\item \link{Boolean}
\item \link{Any}
\item \link{Date}
\item \link{Nested}
\item \link{Url}
\item more coming soon
}
}

\section{Usage}{

You can use any supported fields in the \code{fields} object.

You can call the field like \code{fields$character()},
or pass supported arguments like \code{fields$character(data_key = "foobar")}
}

\examples{
fields
fields$character()
}
\keyword{datasets}
