% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-classes.R
\name{Integer}
\alias{Integer}
\title{Integer}
\description{
A Integer field
}
\keyword{internal}
\section{Super classes}{
\code{\link[staypuft:FieldABC]{staypuft::FieldABC}} -> \code{\link[staypuft:Field]{staypuft::Field}} -> \code{\link[staypuft:Number]{staypuft::Number}} -> \code{Integer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Integer$new()}}
\item \href{#method-format_num}{\code{Integer$format_num()}}
\item \href{#method-clone}{\code{Integer$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="bind_to_schema">}\href{../../staypuft/html/Field.html#method-bind_to_schema}{\code{staypuft::Field$bind_to_schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="context">}\href{../../staypuft/html/Field.html#method-context}{\code{staypuft::Field$context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="deserialize">}\href{../../staypuft/html/Field.html#method-deserialize}{\code{staypuft::Field$deserialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="fail">}\href{../../staypuft/html/Field.html#method-fail}{\code{staypuft::Field$fail()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="get_value">}\href{../../staypuft/html/Field.html#method-get_value}{\code{staypuft::Field$get_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="print">}\href{../../staypuft/html/Field.html#method-print}{\code{staypuft::Field$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="root">}\href{../../staypuft/html/Field.html#method-root}{\code{staypuft::Field$root()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="serialize">}\href{../../staypuft/html/Field.html#method-serialize}{\code{staypuft::Field$serialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="validate_">}\href{../../staypuft/html/Field.html#method-validate_}{\code{staypuft::Field$validate_()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="validate_missing_">}\href{../../staypuft/html/Field.html#method-validate_missing_}{\code{staypuft::Field$validate_missing_()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Number" data-id="deserialize_">}\href{../../staypuft/html/Number.html#method-deserialize_}{\code{staypuft::Number$deserialize_()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Number" data-id="serialize_">}\href{../../staypuft/html/Number.html#method-serialize_}{\code{staypuft::Number$serialize_()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Number" data-id="to_string">}\href{../../staypuft/html/Number.html#method-to_string}{\code{staypuft::Number$to_string()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Number" data-id="validated">}\href{../../staypuft/html/Number.html#method-validated}{\code{staypuft::Number$validated()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Integer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integer$new(..., strict = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strict}}{If \code{TRUE}, only integer types are valid. Otherwise,
any value castable to \code{integer} is valid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_num"></a>}}
\if{latex}{\out{\hypertarget{method-format_num}{}}}
\subsection{Method \code{format_num()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integer$format_num(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
