% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-classes.R
\name{Nested}
\alias{Nested}
\title{Nested}
\description{
Nested

Nested
}
\examples{
artist_schema <- Schema$new("ArtistSchema",
  name = fields$character()
)
x <- Nested$new(artist_schema)
x
x$nested
x$deserialize_(value = list(name = 6)) # good
x$deserialize_(value = list(foobar = 6)) # bad
}
\section{Super classes}{
\code{\link[staypuft:FieldABC]{staypuft::FieldABC}} -> \code{\link[staypuft:Field]{staypuft::Field}} -> \code{Nested}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Nested$new()}}
\item \href{#method-schema}{\code{Nested$schema()}}
\item \href{#method-serialize_}{\code{Nested$serialize_()}}
\item \href{#method-test_list}{\code{Nested$test_list()}}
\item \href{#method-load_}{\code{Nested$load_()}}
\item \href{#method-deserialize_}{\code{Nested$deserialize_()}}
\item \href{#method-clone}{\code{Nested$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="bind_to_schema">}\href{../../staypuft/html/Field.html#method-bind_to_schema}{\code{staypuft::Field$bind_to_schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="context">}\href{../../staypuft/html/Field.html#method-context}{\code{staypuft::Field$context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="deserialize">}\href{../../staypuft/html/Field.html#method-deserialize}{\code{staypuft::Field$deserialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="fail">}\href{../../staypuft/html/Field.html#method-fail}{\code{staypuft::Field$fail()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="get_value">}\href{../../staypuft/html/Field.html#method-get_value}{\code{staypuft::Field$get_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="print">}\href{../../staypuft/html/Field.html#method-print}{\code{staypuft::Field$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="root">}\href{../../staypuft/html/Field.html#method-root}{\code{staypuft::Field$root()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="serialize">}\href{../../staypuft/html/Field.html#method-serialize}{\code{staypuft::Field$serialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="validate_">}\href{../../staypuft/html/Field.html#method-validate_}{\code{staypuft::Field$validate_()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="staypuft" data-topic="Field" data-id="validate_missing_">}\href{../../staypuft/html/Field.html#method-validate_missing_}{\code{staypuft::Field$validate_missing_()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nested$new(
  nested,
  default = miss_ing,
  exclude = NULL,
  only = NULL,
  many = FALSE,
  unknown = "raise",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-schema"></a>}}
\if{latex}{\out{\hypertarget{method-schema}{}}}
\subsection{Method \code{schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nested$schema()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize_"></a>}}
\if{latex}{\out{\hypertarget{method-serialize_}{}}}
\subsection{Method \code{serialize_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nested$serialize_(value, nested_obj, attr = NULL, obj = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_list"></a>}}
\if{latex}{\out{\hypertarget{method-test_list}{}}}
\subsection{Method \code{test_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nested$test_list(self, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_"></a>}}
\if{latex}{\out{\hypertarget{method-load_}{}}}
\subsection{Method \code{load_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nested$load_(value, data = NULL, partial = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserialize_"></a>}}
\if{latex}{\out{\hypertarget{method-deserialize_}{}}}
\subsection{Method \code{deserialize_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nested$deserialize_(value, attr = NULL, data = NULL, partial = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nested$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
