% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puft_fields.R
\docType{data}
\name{puft_fields}
\alias{puft_fields}
\title{staypuft fields}
\format{An object of class \code{list} of length 8.}
\usage{
puft_fields
}
\description{
staypuft fields
}
\details{
types of fields:
\itemize{
\item \code{Missing}
\item \code{Character}
\item \code{UUID}
\item \code{Number}
\item \code{Integer}
\item \code{Boolean}
\item \code{Any}
\item more coming soon ...
}
}
\examples{
puft_fields
puft_fields$field()
puft_fields$missing()
puft_fields$character()
puft_fields$number()
puft_fields$integer()
puft_fields$uuid()
puft_fields$boolean()
puft_fields$any()
}
\keyword{datasets}
