% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schema.R
\docType{data}
\name{Schema}
\alias{Schema}
\title{Schema}
\description{
Schema
}
\details{
\strong{Methods}
\describe{
\item{\code{dump(x, ...)}}{
xxx
}
\item{\code{dump_json(x, ...)}}{
xxx
}
\item{\code{load(x, ...)}}{
xxx
}
\item{\code{load_json(x, ...)}}{
xxx
}
}
}
\examples{
z <- Schema$new("FooBar",
  name = Character$new(),
  title = Character$new()
)
z
z$fields
names(z$fields)

x <- list(name = "Jane Doe", title = "Howdy doody")
x
z$dump(x)
z$dump_json(x)
z$dump_json(x, auto_unbox = TRUE)


z <- Schema$new("MySchema",
  name = Character$new(),
  title = Character$new()
)
z
x <- list(name = "Jane Doe", title = "Howdy doody")
z$load(x)
z$load_json(jsonlite::toJSON(x, auto_unbox=TRUE))

# unknown field
# x <- list(name = "Jane Doe", my_title = "Howdy doody")
# z$load(x)
# z$load_json(jsonlite::toJSON(x, auto_unbox=TRUE))
}
\keyword{datasets}
