% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_field_schema.R
\name{SchemaABC}
\alias{SchemaABC}
\title{SchemaABC}
\description{
Abstract base class from which all Schemas inherit
}
\examples{
\dontrun{
x <- SchemaABC$new()
x
# x$dump()
# x$dump_json()
# x$load()
# x$load_json()
}
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dump}{\code{SchemaABC$dump()}}
\item \href{#method-dump_json}{\code{SchemaABC$dump_json()}}
\item \href{#method-load}{\code{SchemaABC$load()}}
\item \href{#method-load_json}{\code{SchemaABC$load_json()}}
\item \href{#method-clone}{\code{SchemaABC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump"></a>}}
\if{latex}{\out{\hypertarget{method-dump}{}}}
\subsection{Method \code{dump()}}{
dump fun: to be replaced by child class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaABC$dump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump_json"></a>}}
\if{latex}{\out{\hypertarget{method-dump_json}{}}}
\subsection{Method \code{dump_json()}}{
dump_json fun: to be replaced by child class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaABC$dump_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
load fun: to be replaced by child class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaABC$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_json"></a>}}
\if{latex}{\out{\hypertarget{method-load_json}{}}}
\subsection{Method \code{load_json()}}{
load_json fun: to be replaced by child class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaABC$load_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaABC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
