% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staypuft-package.R
\docType{package}
\name{staypuft-package}
\alias{staypuft-package}
\alias{staypuft}
\title{Easy http}
\description{
Easy http
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
