% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\docType{data}
\name{staypuft_fields}
\alias{staypuft_fields}
\alias{Missing}
\alias{miss_ing}
\alias{Character}
\title{staypuft fields}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Missing

miss_ing

Character
}
\description{
staypuft fields
}
\details{
types of fields:
\itemize{
\item \code{Missing}
\item \code{Character}
\item more coming ...
}
}
\keyword{datasets}
