% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\docType{data}
\name{fields}
\alias{fields}
\title{staypuft fields}
\format{
An object of class \code{list} of length 10.
}
\usage{
fields
}
\description{
staypuft fields
}
\details{
types of fields:
\itemize{
\item \code{Missing}
\item \code{Character}
\item \code{UUID}
\item \code{Number}
\item \code{Integer}
\item \code{Boolean}
\item \code{Any}
\item more coming soon
}
}
\examples{
fields
fields$field()
fields$missing()
fields$character()
fields$number()
fields$integer()
fields$uuid()
fields$boolean()
fields$any()
fields$date()
}
\keyword{datasets}
