% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schema.R
\name{Schema}
\alias{Schema}
\title{Schema}
\description{
Base schema class with which to define custom schemas
}
\examples{
z <- Schema$new("FooBar",
  name = puft_fields$character(),
  title = puft_fields$character()
)
z
z$fields
names(z$fields)

x <- list(name = "Jane Doe", title = "Howdy doody")
x
z$dump(x)
z$dump_json(x)
z$dump_json(x, auto_unbox = TRUE)


z <- Schema$new("MySchema",
  name = puft_fields$character(),
  title = puft_fields$character()
)
z
x <- list(name = "Jane Doe", title = "Howdy doody")
z$load(x)
z$load_json(jsonlite::toJSON(x, auto_unbox=TRUE))

# unknown field
# x <- list(name = "Jane Doe", my_title = "Howdy doody")
# z$load(x)
# z$load_json(jsonlite::toJSON(x, auto_unbox=TRUE))

# as data.frame
z <- Schema$new("MySchema",
  name = puft_fields$character(),
  title = puft_fields$character()
)
x <- list(name = "Jane Doe", title = "hello world")
z$load(x, as_df = TRUE)

# list of lists
z <- Schema$new("MySchema",
  name = puft_fields$character(),
  title = puft_fields$character()
)
x <- list(
  list(name = "Jane Doe", title = "hello world"),
  list(name = "Alice Water", title = "bye mars")
)
z$load(x, many = TRUE)
# z$load(x, many = FALSE)

# data.frame's
x <- data.frame(name = "jill", title = "boss", stringsAsFactors = FALSE)
x2 <- data.frame(name = c("jill", "jane"), title = c("boss", "ceo"),
  stringsAsFactors = FALSE)
x2 <- data.frame(name = c("jill", "jane"), title = c("boss", "ceo"),
  stringsAsFactors = FALSE)
z <- Schema$new("FooBar",
  name = puft_fields$character(),
  title = puft_fields$character()
)
z$load_df(x)
z$load_df(x2)
z$load_df(x2, many = TRUE, simplifyVector = FALSE)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{schema_name}}{the schema name}

\item{\code{fields}}{field names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Schema$new()}}
\item \href{#method-print}{\code{Schema$print()}}
\item \href{#method-dump}{\code{Schema$dump()}}
\item \href{#method-dump_json}{\code{Schema$dump_json()}}
\item \href{#method-load}{\code{Schema$load()}}
\item \href{#method-load_json}{\code{Schema$load_json()}}
\item \href{#method-load_df}{\code{Schema$load_df()}}
\item \href{#method-clone}{\code{Schema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Schema} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$new(schema_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{schema_name}}{(character) the schema name}

\item{\code{...}}{additional arguments, passed to \code{fields}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Schema} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
print method for \code{Schema} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump"></a>}}
\subsection{Method \code{dump()}}{
Convert various objects to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$dump(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump_json"></a>}}
\subsection{Method \code{dump_json()}}{
Same as \code{dump()}, but returns JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$dump_json(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{input}

\item{\code{...}}{additional params passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JSON (character)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\subsection{Method \code{load()}}{
Load data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$load(
  data,
  many = FALSE,
  partial = FALSE,
  unknown = "raise",
  as_df = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a named list}

\item{\code{many}}{(logical) Should be set to \code{TRUE} if \code{obj} is a list of
lists. default: \code{FALSE}}

\item{\code{partial}}{(logical) not implemented yet}

\item{\code{unknown}}{(character) one or "raise", "exclude", or "include".
default: "raise"}

\item{\code{as_df}}{(logical) convert to tibble? default: \code{FALSE}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
xxxx
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_json"></a>}}
\subsection{Method \code{load_json()}}{
Same as \code{load()}, but takes JSON as input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$load_json(data, many = FALSE, partial = FALSE, unknown = "raise", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a named list}

\item{\code{many}}{(logical) Should be set to \code{TRUE} if \code{obj} is a list of
lists. default: \code{FALSE}}

\item{\code{partial}}{(logical) not implemented yet}

\item{\code{unknown}}{(character) one or "raise", "exclude", or "include".
default: "raise"}

\item{\code{...}}{additional params passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_df"></a>}}
\subsection{Method \code{load_df()}}{
Same as \code{load()}, but takes a data.frame as input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$load_df(data, many = FALSE, partial = FALSE, unknown = "raise", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a data.frame}

\item{\code{many}}{(logical) Should be set to \code{TRUE} if \code{obj} is a list of
lists. default: \code{FALSE}}

\item{\code{partial}}{(logical) not implemented yet}

\item{\code{unknown}}{(character) one or "raise", "exclude", or "include".
default: "raise"}

\item{\code{...}}{additional params passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
