% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staypuft-package.R
\docType{package}
\name{staypuft-package}
\alias{staypuft-package}
\alias{staypuft}
\title{Convert Complex Objects to and from R Data Structures}
\description{
Convert Complex Objects to and from R Data Structures
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
